UK scrapers with Detailed Notes
This has all the basic API calls for the entire UK power grid minus the Isle of Man. It includes detailed notes about the type of data received and the frequency they update at. This will need to be processed and set to run at a schedule to be complete