Portugal Scraper
This adds a Portugal scraper based on this api https://e-redes.opendatasoft.com/explore/dataset/outages-per-geography/api/. It is important to note that this data is not as complete as other sources. It may just be which geographic areas are experiencing a current outage rather than any metric about the severity of that outage.